pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "biblioUtils"
description = "Utility scripts for the management of the bibliography of Philosophie.ch"
version = "2024.0.1"
requires-python = ">=3.11.0"
readme = "README.md"
authors = [
  {name = "Luis Bordo", email = "luis.bordo@philosophie.ch"},
]
maintainers = [
  {name = "Luis Bordo", email = "luis.bordo@philosophie.ch"},
]

dependencies = [
  "titlecase>=2.4,<3.0",
  "pydantic>=2.7,<3.0",
  "requests", 
  "fuzzywuzzy",
  "python-Levenshtein",
  "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1,<9.0",
    "mypy>=1.10,<2.0",
    "black>=24.4,<25.0",
]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "titlecase",
    "titlecase_constants",
    "fuzzywuzzy",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -s"
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''